feat: backend finish

+++ controllers/card.js
@@ -4,21 +4,21 @@ const {
   DEFAULT_ERROR_CODE,
   INCORRECT_DATA_ERROR_CODE,
 } = require('../utils/constants');
+const BadRequestError = require("../utils/errors/badRequestError");
+const NotFoundError = require("../utils/errors/notFoundError");
 
 const USER_REF = [{ path: 'likes', model: 'user' }];
 
-module.exports.getCards = async (req, res) => {
+module.exports.getCards = async (req, res, next) => {
   try {
     const cards = await Card.find({});
     res.send(cards);
   } catch (e) {
-    res.status(DEFAULT_ERROR_CODE).json({
-      message: 'Не удалось получить карточки',
-    });
+    next(e);
   }
 };
 
-module.exports.createCard = async (req, res) => {
+module.exports.createCard = async (req, res, next) => {
   try {
     const { name, link } = req.body;
     const card = await Card.create({ name, link, owner: req.user._id });
@@ -28,23 +28,23 @@ module.exports.createCard = async (req, res) => {
     });
   } catch (e) {
     if (e.name === 'ValidationError') {
-      res.status(INCORRECT_DATA_ERROR_CODE).json({
-        message: 'Переданы не валидные данные'
-      });
-      return;
+      next(new BadRequestError(e.message));
     }
-    res.status(DEFAULT_ERROR_CODE).json({
-      message: 'Не удалось создать карточку',
-    });
+    next(e);
   }
 };
 
-module.exports.deleteCard = async (req, res) => {
+module.exports.deleteCard = async (req, res, next) => {
   try {
     const card = await Card.findById(req.params.cardId);
     if (!card) {
       return res.status(NOT_FOUND_ERROR_CODE).json({
         message: 'Карточка не найдена',
+      });
+    }
+    if (card.owner.toString() !== req.params.cardId) {
+      return res.status(NOT_FOUND_ERROR_CODE).json({
+        message: 'Нельзя удалять чужие карточки',
       });
     }
     const cardDelete = await Card.findByIdAndRemove(req.params.cardId);
@@ -53,18 +53,13 @@ module.exports.deleteCard = async (req, res) => {
     });
   } catch (e) {
     if (e.name === 'CastError') {
-      res.status(INCORRECT_DATA_ERROR_CODE).json({
-        message: 'Переданы не валидные данные'
-      });
-      return;
+      next(new BadRequestError('Переданы не валидные данные'));
     }
-    res.status(DEFAULT_ERROR_CODE).json({
-      message: 'Не удалось удалить карточку',
-    });
+    next(e);
   }
 };
 
-const handleCardLike = async (req, res, options) => {
+const handleCardLike = async (req, res, next, options) => {
   try {
     const action = options.addLike ? '$addToSet' : '$pull';
 
@@ -83,14 +78,9 @@ const handleCardLike = async (req, res, options) => {
     res.send(updatedCard);
   } catch (e) {
     if (e.name === 'CastError') {
-      res.status(INCORRECT_DATA_ERROR_CODE).json({
-        message: 'Переданы не валидные данные'
-      });
-      return;
+      next(new BadRequestError('Переданы не валидные данные'));
     }
-    res.status(DEFAULT_ERROR_CODE).json({
-      message: 'Не удалось изменить карточку',
-    });
+    next(e);
   }
 };
